Client-side time shifting support (in-memory only).
This implementation provides fast forward and rewind features.
Currently only in-memory buffer is implemented but it's intended to be
extended with an on disk implementation. The buffer is allocated on the
client side (in the XMBC addon), this approach reduces server loading,
especially, in case of multiple clients.

By default time shifting is disabled (buffer size is 0), thus, it must
not impact on the current functionality.